pyproject.toml
[tool.poetry]
name = "oidconv"
version = "0.1.0"
description = ""
authors = ["naomori <naoki.morita@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.1.4"
Pillow = "^8.0.1"
more-itertools = "^8.6.0"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
